Move a few basic enum types from lyon_tessellation to lyon_path.
This will allow other algorithms that want to refer to line joins or caps to not depend on lyon_tessellation or duplicate the types.
